Extract native-conversion-ish functions into `native-conversion.ts`.
This also drops the `Modern` suffix/infix from all the names, and
stops doing nominal-dispatch in `fabric-value.ts` instead just
re-exporting the functions in question.
